requirements.txt
telethon
python-dotenv
requests
unidecode
asyncio